requirements.txt
appnope==0.1.0
decorator==4.1.2
dj-database-url==0.4.2
dj-static==0.0.6
Django==1.11.5
django-extensions==1.9.1
django-jet==1.0.6
gunicorn==19.7.1
ipython==6.2.1
ipython-genutils==0.2.0
jedi==0.11.0
parso==0.1.0
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
psycopg2==2.7.3.1
ptyprocess==0.5.2
Pygments==1.6
python-decouple==3.1
pytz==2017.2
simplegeneric==0.8.1
six==1.11.0
sqlformatter==1.3
sqlparse==0.1.11
static3==0.7.0
traitlets==4.3.2
wcwidth==0.1.7
django-import-export==0.5.1
